Simplify build and enable back-publish support for new Scala versions (#28)
* Refactored the build. The `plugin` cross-module is no longer a
sub-module of the `moduledefs` cross-module. This removes the need to
specify the correct Scala version of the `moduledefs` module. For
example, to build the plugin for Scala `3.8.0-RC3`, it is now enough to
run `mill plugin[3.8.0-RC3].compile`, whereas before, we needed to run
`mill moduledefs[3.7.4].plugin[3.8.0-RC3].compile`. The artifact IDs
weren't changed.

* When a `SCALA_VERSION` env variable is defined, that Scala version is
added to the target versions to which the plugin will be build. Example:
To build the plugin for Scala `3.8.0-RC4` without changing the
`build.mill` file, you can now run `SCALA_VERSION=3.8.0-RC4 mill
plugin[3.8.0-RC4].compile`.

* The Github publish workflow dispatch now supports a `SCALA_VERSION`
parameter. When set, the workflow will only publish the plugin for that
specific Scala version. This allows us to back-publish for a specific
Scala version without any changes to the repository. (The build workflow
will still build all artifacts.) It is still preferred, to open a PR to
add a new Scala version permanently. The new back-publishing allows us
to support Scala updates in projects that want to stay on an older Mill
version.

* Removed the `publishToSonatype` command from the `build.mill` file,
which was no longer in use and looked outdated, as it was not using the
newer `SonatypeCentralPublishModule`.

* Updated the developer instructions and deleted mentions of the
previous back-publish workflow.

Pull request: https://github.com/com-lihaoyi/mill-moduledefs/pull/28